Updated MacOS X support, including support for mx_is_valid_pointer() using vm_region().
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1226 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_bluice_area_detector.c
@@ -281,13 +281,13 @@ mxd_bluice_area_detector_collect_thread( MX_THREAD *thread, void *args )
 	snprintf( start_oscillation_format,
 		sizeof(start_oscillation_format),
 			"%%%ds %%%ds %%lg",
-			sizeof(request_name)+1,
-			sizeof(motor_name)+1 );
+			(int) sizeof(request_name)+1,
+			(int) sizeof(motor_name)+1 );
 
 	snprintf( prepare_for_oscillation_format,
 		sizeof(prepare_for_oscillation_format),
 			"%%%ds %%lg",
-			sizeof(request_name)+1 );
+			(int) sizeof(request_name)+1 );
 
 	/*-------------------------------------------*/
